fix(playwright): close cross-file race in shared time-range lineage seeding
v3-lineage-impact-analysis.spec.ts's "can see when the inputs to a data
job change" failed under workers_per_shard=2 (run 31360954110, shard 2):
expected user_profile hidden in the 14-to-7-days-ago window, got visible.

Root cause: seedTimeRangeLineage() is called from test.beforeAll in TWO
separate spec files (this one and v3-lineage-graph.spec.ts), each on the
undocumented assumption that beforeAll runs exactly once per full test
run. That's only true under workers:1. With fullyParallel scheduling and
workers_per_shard > 1, Playwright can run these two files' beforeAll
hooks concurrently on different workers. seedTimeRangeLineage does a
delete-then-recreate of shared dataJobInputOutput/upstreamLineage aspects
per its own comments, deletes and the graph-edge writes they gate on
propagate asynchronously, so interleaving two independent delete/recreate
sequences can leave a wrong-timestamp result even though each sequence
alone is correct. Same bug class as the global-data seeding race fixed
in bb4c65e47d, just reached through fullyParallel file-level scheduling
instead of workers_per_shard's worker-level scheduling.

Fix:
- Extracted withArtifactLock/writeJsonAtomic out of seeding.fixture.ts
  into helpers/seeder-utils.ts (cubic flagged the original as effectively
  single-use scaffolding worth sharing; this is the second real caller).
- seedTimeRangeLineage() now runs its body inside withArtifactLock keyed
  to .seeded/lineage-time-range.json, so only one worker across the
  whole run -- regardless of which file's beforeAll gets there first --
  actually performs the seed.
- Added mode:'serial' to both spec files as defensive-in-depth: it makes
  each file's own beforeAll genuinely single-shot regardless of the lock,
  and covers v3-lineage-graph.spec.ts's separate seedAdditionalNodes call
  (unique to that file, not covered by the shared lock).
