+++ e2e-test/ui/playwright/tests/lineage-v3/v3-lineage-graph.spec.ts
@@ -233,6 +233,14 @@ const BASE_FEATURE_FLAGS = {
 
 // ── Test Suite ───────────────────────────────────────────────────────────────
 
+// beforeAll's seedAdditionalNodes call (unlike seedTimeRangeLineage, which is
+// now lock-protected in lineage-time-seeder.ts because it's also called from
+// v3-lineage-impact-analysis.spec.ts) is unique to this file and has no such
+// guard. Under workers_per_shard > 1, fullyParallel can still schedule this
+// describe block's own tests across two workers, each running beforeAll
+// concurrently. Force one worker so it stays genuinely single-shot.
+test.describe.configure({ mode: 'serial' });
+
 test.describe('lineage v3 — lineage graph', () => {
   let lineagePage: LineageV3Page;
