+++ e2e-test/ui/playwright/utils/lineage-time-seeder.ts
@@ -25,9 +25,13 @@
  *     factor_income → gnp  (created 8 days ago, updated 8 days ago — removed since then)
  */
 
+import * as path from 'path';
 import type { APIRequestContext } from '@playwright/test';
 import { gmsUrl } from './constants';
 import type { DataHubLogger } from './logger';
+import { withArtifactLock, writeJsonAtomic } from '../helpers/seeder-utils';
+
+const STATE_FILE = path.join(__dirname, '../.seeded/lineage-time-range.json');
 
 // ── Timestamp helpers ─────────────────────────────────────────────────────────
 
@@ -199,10 +203,34 @@ async function ingestProposal(
 // ── Public seeding entry point ────────────────────────────────────────────────
 
 /**
- * Seed all time-range lineage scenarios.
- * Safe to call multiple times — each call simply overwrites the aspect.
+ * Seed all time-range lineage scenarios, guarded by withArtifactLock so it
+ * only actually runs once across the whole test run.
+ *
+ * Two different spec files (v3-lineage-impact-analysis.spec.ts and
+ * v3-lineage-graph.spec.ts) each call this from their own test.beforeAll,
+ * on the — previously correct only under workers:1 — assumption that
+ * "once per describe block" means "once, period". Under workers_per_shard
+ * > 1 (or just fullyParallel scheduling these two files onto different
+ * workers), both beforeAll hooks can run concurrently, each doing its own
+ * delete-then-recreate of the SAME shared dataJobInputOutput/upstreamLineage
+ * aspects (see doSeedTimeRangeLineage). Deletes and the graph-edge writes
+ * they gate on propagate asynchronously, so interleaving two independent
+ * delete/recreate sequences can leave a partially-overwritten, wrong-
+ * timestamp result even though each sequence alone is correct -- exactly
+ * the class of bug the seeding fixture's global-data race was.
  */
 export async function seedTimeRangeLineage(
+  request: APIRequestContext,
+  gmsToken: string,
+  logger?: DataHubLogger,
+): Promise<void> {
+  await withArtifactLock(STATE_FILE, async () => {
+    await doSeedTimeRangeLineage(request, gmsToken, logger);
+    writeJsonAtomic(STATE_FILE, { seededAt: new Date().toISOString() });
+  });
+}
+
+async function doSeedTimeRangeLineage(
   request: APIRequestContext,
   gmsToken: string,
   logger?: DataHubLogger,
